fix: add generic support
